requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==3.13.0
django-filter==22.1
django-multiupload==0.6.1
django-picklefield==3.1
django-templated-mail==1.1.1
django-countries==7.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
idna==3.4
imutils==0.5.4
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
mypy-extensions==1.0.0
numpy==1.24.1
oauthlib==3.2.2
packaging==22.0
pathspec==0.11.0
Pillow==9.3.0
platformdirs==3.0.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sensecam-discovery==2.2.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
zeep==4.2.1
# ms sql connector
pyodbc==4.0.35
async-timeout==4.0.2
# redis packages
redis==4.5.4
django-redis==5.2.0
# celery packages
django-crontab==0.7.1
schedule==1.2.0
celery==5.2.7
django-extensions==3.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.0
psycopg2-binary==2.9.6
